pyproject.toml
[project]
name = "steam-analytics-fabric"
version = "0.1.0"
description = "Enterprise-grade data pipeline for Steam analytics using Microsoft Fabric"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Eduardo Jafet Cendon Aguilar", email = "eduardo.jafet31oct@gmail.com" }
]
keywords = ["steam", "analytics", "fabric", "data-engineering", "medallion"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "httpx>=0.27.0",              # HTTP client async-first
    "pydantic>=2.5.0",            # Data contracts/validation
    "pydantic-settings>=2.1.0",   # Config management
    "tenacity>=8.2.0",            # Retry logic
    "structlog>=24.1.0",          # Structured logging
    "python-dotenv>=1.0.0",       # Environment variables
    "azure-identity>=1.15.0",
    "azure-storage-file-datalake>=12.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",                # Linting + formatting
    "mypy>=1.8.0",                # Type checking
    "pre-commit>=3.6.0",          # Git hooks
    "respx>=0.20.0",              # Mock HTTP para tests
]
quality = [
    "great-expectations>=0.18.0",
    "pandera>=0.18.0",
]
spark = [
    "pyspark>=3.5.0",
    "delta-spark>=3.1.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
]

[project.urls]
Homepage = "https://github.com/gatotroller/steam-analytics-fabric"
Documentation = "https://github.com/gatotroller/steam-analytics-fabric#readme"
Repository = "https://github.com/gatotroller/steam-analytics-fabric.git"
Issues = "https://github.com/gatotroller/steam-analytics-fabric/issues"

[tool.ruff]
target-version = "py310"
line-length = 100
exclude = [".venv", "notebooks"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
strict = true
python_version = "3.10"
ignore_missing_imports = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
plugins = ["pydantic.mypy"]
files = ["src"]
explicit_package_bases = true
namespace_packages = true
disallow_untyped_decorators = false
disallow_subclassing_any = false

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]
addopts = [
    "-v",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-fail-under=50",
]

[tool.coverage.run]
source = ["src"]
branch = true
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]

[tool.hatch.build.targets.wheel]
packages = ["src/steam_analytics"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
